deterministic route names and test_component_from_yaml
