+++ pp/components/bezier.py
@@ -1,4 +1,3 @@
-import hashlib
 from typing import List, Optional, Tuple
 
 import numpy as np
@@ -9,6 +8,7 @@ from scipy.special import binom
 import pp
 from pp.component import Component
 from pp.geo_utils import angles_deg, curvature, extrude_path, path_length, snap_angle
+from pp.hash_points import hash_points
 from pp.layers import LAYER
 
 
@@ -63,14 +63,8 @@ def bezier(
         layer: layer
     """
 
-    def format_float(x):
-        return "{:.3f}".format(x).rstrip("0").rstrip(".")
-
-    def _fmt_cp(cps):
-        return "_".join([f"({format_float(p[0])},{format_float(p[1])})" for p in cps])
-
     if name is None:
-        points_hash = hashlib.md5(_fmt_cp(control_points).encode()).hexdigest()
+        points_hash = hash_points(control_points)
         name = f"bezier_w{int(width*1e3)}_{points_hash}_{layer[0]}_{layer[1]}"
 
     c = pp.Component(name=name)
