+++ pp/test_component_from_yaml/test_ports_sample_mmis_.csv
@@ -1,0 +1,7 @@
+,E0,W0
+0,-12.5,115
+1,0,100
+2,180,0
+3,0.5,0.5
+4,1,1
+5,0,0
