+++ pp/test_component_from_yaml/test_settings_sample_different_factory_.yml
@@ -1,0 +1,19 @@
+info:
+  instances:
+    bl: DeviceReference (parent Device "pad", ports ['N', 'S', 'E', 'W'], origin (0,
+      0), rotation 0, x_reflection False)
+    br: DeviceReference (parent Device "pad", ports ['N', 'S', 'E', 'W'], origin [400.
+      400.], rotation 0, x_reflection False)
+    tl: DeviceReference (parent Device "pad", ports ['N', 'S', 'E', 'W'], origin [  0.
+      200.], rotation 0, x_reflection False)
+    tr: DeviceReference (parent Device "pad", ports ['N', 'S', 'E', 'W'], origin [400.
+      600.], rotation 0, x_reflection False)
+  routes:
+    bl,E:br,W: DeviceReference (parent Device "zz_conn_6be4ce7aff81eacc", ports ['input',
+      'output'], origin [0 0], rotation 0, x_reflection False)
+    bl,S:br,E: DeviceReference (parent Device "zz_conn_d26997c15fa0e68d", ports ['input',
+      'output'], origin [0 0], rotation 0, x_reflection False)
+    tl,E:tr,W: DeviceReference (parent Device "zz_conn_73ebe2ba20581b39", ports ['input',
+      'output'], origin [0 0], rotation 0, x_reflection False)
+name: sample_different_factory
+settings: {}
