+++ pp/test_component_from_yaml/test_settings_sample_different_link_factory_.yml
@@ -1,0 +1,17 @@
+info:
+  instances:
+    bl: DeviceReference (parent Device "pad", ports ['N', 'S', 'E', 'W'], origin (0,
+      0), rotation 0, x_reflection False)
+    br: DeviceReference (parent Device "pad", ports ['N', 'S', 'E', 'W'], origin [900.
+      400.], rotation 0, x_reflection False)
+    tl: DeviceReference (parent Device "pad", ports ['N', 'S', 'E', 'W'], origin [  0.
+      200.], rotation 0, x_reflection False)
+    tr: DeviceReference (parent Device "pad", ports ['N', 'S', 'E', 'W'], origin [900.
+      600.], rotation 0, x_reflection False)
+  routes:
+    bl,E:br,W: DeviceReference (parent Device "zz_conn_11ab64a88b4704e8", ports ['input',
+      'output'], origin [0 0], rotation 0, x_reflection False)
+    tl,E:tr,W: DeviceReference (parent Device "zz_conn_510d926c0c150d44", ports ['input',
+      'output'], origin [0 0], rotation 0, x_reflection False)
+name: sample_different_link_factory
+settings: {}
