+++ pp/test_component_from_yaml/test_settings_sample_docstring_.yml
@@ -1,0 +1,11 @@
+info:
+  instances:
+    mmi_bot: DeviceReference (parent Device "mmi1x2_LM10_WM4p5", ports ['E0', 'E1',
+      'W0'], origin [ 70.    -29.375], rotation 0, x_reflection False)
+    mmi_top: DeviceReference (parent Device "mmi1x2_LM5_WM4p5", ports ['E0', 'E1',
+      'W0'], origin [12.5  0. ], rotation 0, x_reflection False)
+  routes:
+    mmi_top,E0:mmi_bot,W0: DeviceReference (parent Device "zz_conn_5baf95cc396bb19d",
+      ports ['input', 'output'], origin [0 0], rotation 0, x_reflection False)
+name: sample_docstring
+settings: {}
