+++ pp/test_component_from_yaml/test_settings_sample_regex_connections_backwards_.yml
@@ -1,0 +1,15 @@
+info:
+  instances:
+    left: DeviceReference (parent Device "nxn_E3_W0_Y20", ports ['E0', 'E1', 'E2'],
+      origin (0, 0), rotation 0, x_reflection False)
+    right: DeviceReference (parent Device "nxn_E0_W3_Y20", ports ['W0', 'W1', 'W2'],
+      origin [20.  0.], rotation 0, x_reflection False)
+  routes:
+    left,E0:right,W0: DeviceReference (parent Device "zz_conn_db5622e16984eeca", ports
+      ['input', 'output'], origin [0 0], rotation 0, x_reflection False)
+    left,E1:right,W1: DeviceReference (parent Device "zz_conn_9325a15c05a3909c", ports
+      ['input', 'output'], origin [0 0], rotation 0, x_reflection False)
+    left,E2:right,W2: DeviceReference (parent Device "zz_conn_bd78140a553112b6", ports
+      ['input', 'output'], origin [0 0], rotation 0, x_reflection False)
+name: sample_regex_connections_backwards
+settings: {}
