+++ pp/test_component_from_yaml/test_settings_sample_waypoints_.yml
@@ -1,0 +1,11 @@
+info:
+  instances:
+    b: DeviceReference (parent Device "pad_array", ports ['N0', 'N1', 'N2', 'N3',
+      'N4', 'N5'], origin (0, 0), rotation 0, x_reflection False)
+    t: DeviceReference (parent Device "pad_array_PL'S'", ports ['S0', 'S1', 'S2',
+      'S3', 'S4', 'S5'], origin [ 100. 1000.], rotation 0, x_reflection False)
+  routes:
+    t,S5:b,N4: DeviceReference (parent Device "zz_conn_56119df9aeba8802", ports ['input',
+      'output'], origin [0 0], rotation 0, x_reflection False)
+name: sample_waypoints
+settings: {}
